fix hanging test
